refactor: extract shared settlement openapi routing helpers [mojaloop/#4479]
SonarCloud reported 6.6% duplication on new code (gate requires <= 3%): getBasePath
and handleRequest were byte-identical in src/settlement/api/routes.js and
handlerRoutes.js, 46 duplicated lines each.

Both now import them from src/settlement/api/openapiRouting.js. No behaviour change.
