+++ package.json
@@ -96,7 +96,7 @@
     "@mojaloop/central-services-health": "15.2.3",
     "@mojaloop/central-services-logger": "11.10.6",
     "@mojaloop/central-services-metrics": "12.8.8",
-    "@mojaloop/central-services-shared": "18.38.2",
+    "@mojaloop/central-services-shared": "18.37.0",
     "@mojaloop/central-services-stream": "11.9.3",
     "@mojaloop/database-lib": "11.3.10",
     "@mojaloop/event-sdk": "14.8.5",
@@ -117,7 +117,7 @@
     "hapi-auth-bearer-token": "8.0.0",
     "hapi-swagger": "17.3.2",
     "ilp-packet": "2.2.0",
-    "joi": "18.2.3",
+    "joi": "18.2.5",
     "knex": "3.3.0",
     "lodash": "4.18.1",
     "moment": "2.30.1",
