+++ src/settlement/api/openapiRouting.js
@@ -28,43 +28,38 @@
  ******/
 'use strict'
 
-const OpenapiBackend = require('@mojaloop/central-services-shared').Util.OpenapiBackend
-const Path = require('path')
-const Handlers = require('./handlers')
-const { getBasePath, handleRequest } = require('./openapiRouting')
+/**
+ * Routing helpers shared by the settlement API and the settlement handlers
+ * monitoring API, both of which are served by OpenapiBackend.
+ */
 
 /**
- * Handlers monitoring API Routes
+ * Base path the API is served under. It mirrors the `servers` url of the
+ * OpenAPI document (formerly the Swagger 2.0 `basePath`).
  *
  * @param {object} api OpenAPIBackend instance
+ * @returns {string} base path, e.g. '/v2'
  */
-const APIRoutes = (api) => {
-  const basePath = getBasePath(api)
-  return [
-    {
-      method: 'GET',
-      path: `${basePath}/health`,
-      handler: (req, h) => handleRequest(api, req, h),
-      config: {
-        id: 'handlers getHealth',
-        tags: ['api', 'getHealth'],
-        description: 'GET health'
-      }
-    }
-  ]
-}
+const getBasePath = (api) => api.definition.servers[0].url.replace(/\/$/, '')
+
+/**
+ * Request handler. The base path is stripped from the request path, as the
+ * paths of the OpenAPI document are relative to its `servers` url.
+ *
+ * @param {object} api OpenAPIBackend instance
+ * @param {object} req Request
+ * @param {object} h   Response handle
+ */
+const handleRequest = (api, req, h) => api.handleRequest(
+  {
+    method: req.method,
+    path: req.path.slice(getBasePath(api).length),
+    body: req.payload,
+    query: req.query,
+    headers: req.headers
+  }, req, h)
 
 module.exports = {
-  plugin: {
-    name: 'settlement handler api routes',
-    register: async function (server) {
-      const api = await OpenapiBackend.initialise(Path.resolve(process.cwd(), 'src/settlement/interface/swagger-handler.json'), {
-        getHealth: Handlers.getHealth,
-        validationFail: Handlers.validationFail,
-        notFound: Handlers.notFound,
-        methodNotAllowed: Handlers.methodNotAllowed
-      })
-      server.route(APIRoutes(api))
-    }
-  }
+  getBasePath,
+  handleRequest
 }
