+++ src/shared/plugins.js
@@ -51,7 +51,7 @@ const registerPlugins = async (server) => {
     await server.register({
       plugin: APIDocumentation,
       options: {
-        pathToSwaggerFile: Path.resolve(process.cwd(), 'src/api/interface/swagger.json')
+        documentPath: Path.resolve(process.cwd(), 'src/api/interface/swagger.json')
       }
     })
   }
